AC_Avoidance: Dijkstras split into common, static data and calcpath
